services/cms/requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
psycopg2-binary==2.7.5
pytz==2018.5
redis==2.10.6
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
